Pick Them All pack: rework Blood Wall into Bulk Up, replacing the HP loss with a max HP gain and moving the block and temporary HP back to the previous values
